+++ src/opinf/lstsq/__init__.py
@@ -2,5 +2,6 @@
 r"""Solvers for Operator Inference least-squares problems."""
 
 from ._base import *
+from ._tsvd import *
 from ._tikhonov import *
 from ._total import *