+++ src/opinf/models/mono/_parametric.py
@@ -312,7 +312,7 @@ class _ParametricModel(_Model):
             return self
 
         # Execute non-intrusive learning.
-        self._extract_operators(np.atleast_2d(self.solver.predict()))
+        self._extract_operators(self.solver.solve())
 
     def fit(self, parameters, states, lhs, inputs=None):
         r"""Learn the model operators from data.