+++ tests/lstsq/test_base.py
@@ -47,7 +47,7 @@ class TestSolverTemplate:
     class Dummy(opinf.lstsq._base.SolverTemplate):
         """Instantiable version of SolverTemplate."""
 
-        def predict(self):
+        def solve(self):
             return np.ones((self.r, self.d))
 
     def test_properties(self):
@@ -194,8 +194,9 @@ class TestSolverTemplate:
         assert solver.r == 1
         ohat = np.random.standard_normal(d)
         residual = solver.residual(ohat)
-        assert isinstance(residual, float)
-        assert np.isclose(residual, la.norm(D @ ohat - z) ** 2)
+        assert isinstance(residual, np.ndarray)
+        assert residual.shape == (1,)
+        assert np.isclose(residual[0], la.norm(D @ ohat - z) ** 2)
 
     def test_copy(self, k=18, d=10, r=4):
         """Test copy()."""
@@ -219,11 +220,15 @@ class TestSolverTemplate:
         assert np.all(solver2.lhs_matrix == Z)
 
     # Verification ------------------------------------------------------------
-    def test_verify(self):
+    def test_verify(self, r=5, k=40, d=10):
         """Test verify()."""
+        D = np.random.random((k, d))
+        Z = np.random.random((r, k))
+        self.Dummy().verify()
 
         def _single(DClass, message):
             dummy = DClass()
+            dummy.fit(D, Z)
             with pytest.raises(opinf.errors.VerificationError) as ex:
                 dummy.verify()
             assert ex.value.args[0].startswith(message)
@@ -234,23 +239,11 @@ class TestSolverTemplate:
 
         _single(Dummy2, "Dummy2.copy() returned object of type 'int'")
 
-        class Dummy3(self.Dummy):
-            def fit(self, D, Z):
-                self.Ohat = D - Z
-
-        _single(Dummy3, "fit() failed")
-
-        class Dummy4(self.Dummy):
-            def fit(self, D, Z):
-                pass
-
-        _single(Dummy4, "fit() should call SolverTemplate.fit()")
-
         class Dummy5(self.Dummy):
-            def predict(self):
+            def solve(self):
                 return np.empty((1, 1))
 
-        _single(Dummy5, "predict() did not return array of shape (r, d)")
+        _single(Dummy5, "solve() did not return array of shape (r, d)")
 
         class Dummy6(self.Dummy):
             def copy(self):
@@ -262,10 +255,10 @@ class TestSolverTemplate:
         _single(Dummy6, "copy() does not preserve problem dimensions")
 
         class Dummy7(self.Dummy):
-            def predict(self):
+            def solve(self):
                 return np.random.random((self.r, self.d))
 
-        _single(Dummy7, "copy() does not preserve the result of predict()")
+        _single(Dummy7, "copy() does not preserve the result of solve()")
 
         class Dummy8(self.Dummy):
             def save(self, savefile, overwrite=False):
@@ -281,7 +274,7 @@ class TestSolverTemplate:
 
         _single(Dummy8, "save()/load() does not preserve problem dimensions")
 
-        self.Dummy().verify()
+        self.Dummy().fit(D, Z).verify()
 
 
 class TestPlainSolver:
@@ -316,8 +309,8 @@ class TestPlainSolver:
 
         repr(solver)
 
-    def test_predict(self, k=20, d=11, r=3):
-        """Test predict()."""
+    def test_solve(self, k=20, d=11, r=3):
+        """Test solve()."""
         # Set up and manually solve a least-squares problem.
         D = np.random.standard_normal((k, d))
         Z = np.random.random((r, k))
@@ -326,7 +319,7 @@ class TestPlainSolver:
 
         # Check the least-squares solution.
         solver = self.Solver().fit(D, Z)
-        Ohat = solver.predict()
+        Ohat = solver.solve()
         assert np.allclose(Ohat, Ohat_true)
 
     def test_save(self, k=6, d=4, r=2, outfile="_plainsolversavetest.h5"):